database/requirements.txt
geopandas
pandas
pyarrow
pyyaml
sqlalchemy
tqdm